task_manager/requirements.txt
amqp==5.1.1
billiard==4.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
docker==6.1.3
idna==3.4
kombu==5.3.2
packaging==23.2
prompt-toolkit==3.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
websocket-client==1.6.4
# pywin32==306